+++ DataDefinitions/Station.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Utilities;
 
 namespace EddiDataDefinitions
 {
@@ -83,13 +82,13 @@ namespace EddiDataDefinitions
         public List<Ship> shipyard { get; set; }
 
         // Admin - the last time the information present changed
-        public DateTime? updatedat;
+        public long? updatedat;
 
         // Admin - the last time the market information present changed
-        public DateTime? commoditiesupdatedat;
+        public long? commoditiesupdatedat;
 
         // Admin - the last time the outfitting information present changed
-        public DateTime? outfittingupdatedat;
+        public long? outfittingupdatedat;
 
         /// <summary>Is this station a starport?</summary>
         public bool IsStarport() { return model == null ? false : model.EndsWith("Starport"); }