+++ Utilities/Logging.cs
@@ -52,19 +52,57 @@ namespace Utilities
             log(filePath, memberName, "W", message + " " + data);
         }
 
-        public static void Info(string data, [CallerMemberName] string memberName = "", [CallerFilePath] string filePath = "")
+        public static void Info(string message, [CallerMemberName] string memberName = "", [CallerFilePath] string filePath = "")
         {
-            log(filePath, memberName, "I", data);
+            Info(message, (string)null, memberName, filePath);
         }
 
-        public static void Debug(string data, [CallerMemberName] string memberName = "", [CallerFilePath] string filePath = "")
+        public static void Info(string message, Exception ex, [CallerMemberName] string memberName = "", [CallerFilePath] string filePath = "")
+        {
+            Info(message, ex.ToString(), memberName, filePath);
+        }
+
+        public static void Info(Exception ex, [CallerMemberName] string memberName = "", [CallerFilePath] string filePath = "")
+        {
+            Info(ex.ToString(), memberName, filePath);
+        }
+
+        public static void Info(string message, string data, [CallerMemberName] string memberName = "", [CallerFilePath] string filePath = "")
+        {
+            log(filePath, memberName, "I", message + " " + data);
+        }
+
+        public static void Debug(string message, [CallerMemberName] string memberName = "", [CallerFilePath] string filePath = "")
         {
             if (Verbose)
             {
-                log(filePath, memberName, "D", data);
+                Debug(message, (string)null, memberName, filePath);
             }
         }
 
+        public static void Debug(string message, Exception ex, [CallerMemberName] string memberName = "", [CallerFilePath] string filePath = "")
+        {
+            if (Verbose)
+            {
+                Debug(message, ex.ToString(), memberName, filePath);
+            }
+        }
+
+        public static void Debug(Exception ex, [CallerMemberName] string memberName = "", [CallerFilePath] string filePath = "")
+        {
+            if (Verbose)
+            {
+                Debug(ex.ToString(), memberName, filePath);
+            }
+        }
+
+        public static void Debug(string message, string data, [CallerMemberName] string memberName = "", [CallerFilePath] string filePath = "")
+        {
+            if (Verbose)
+            {
+                log(filePath, memberName, "D", message + " " + data);
+            }
+        }
         private static readonly object logLock = new object();
         private static void log(string path, string method, string level, string data)
         {
@@ -109,7 +147,10 @@ namespace Utilities
                 thread.IsBackground = true;
                 thread.Start();
             }
-            catch { }
+            catch (Exception)
+            {
+                // Nothing to do
+            }
         }
     }
 }